fix: int cast in MusicBrainzApi::applyRateLimiting() and add unit tests

+++ src/MusicBrainzApi.php
@@ -365,6 +365,6 @@ final class MusicBrainzApi
             }
         }
 
-        $this->lastRequestTime = microtime(true) * 1000;
+        $this->lastRequestTime = (int)(microtime(true) * 1000);
     }
 }
